pulled out scroll transition logic into a separate component to make way for fade-transition
