+++ test/specs/carousel.test.js
@@ -286,12 +286,14 @@ describe('<Carousel />', () => {
       const elems = ['Slide 2', 'Slide 3', 'Slide 4'];
       const wrapper = mount(
         <Carousel>
-          <p>Static Slide</p>
+          <p key="Static Slide">Static Slide</p>
           {elems.map(e => `<p key={${e}}>${e}</e>`)}
         </Carousel>
       );
       expect(wrapper).toHaveState({ slideCount: 4 });
-      const children = wrapper.props().children.concat(<p>Slide 4</p>);
+      const children = wrapper
+        .props()
+        .children.concat(<p key="Slide 5">Slide 4</p>);
       wrapper.setProps({ children });
       expect(wrapper).toHaveState({ slideCount: 5 });
     });
@@ -407,58 +409,6 @@ describe('<Carousel />', () => {
       expect(wrapper).toHaveState({ currentSlide: 2 });
       nextButton.simulate('click');
       expect(wrapper).toHaveState({ currentSlide: 0 });
-    });
-
-    describe('#getSlideDirection', () => {
-      let instance;
-
-      beforeEach(async () => {
-        const wrapper = mount(
-          <Carousel>
-            <p>Slide 1</p>
-            <p>Slide 2</p>
-            <p>Slide 3</p>
-          </Carousel>
-        );
-        instance = wrapper.instance();
-      });
-
-      it('should return zero if start and end slide are the same regardless of isWrapping', () => {
-        expect(instance.getSlideDirection(2, 2, true)).toEqual(0);
-        expect(instance.getSlideDirection(2, 2, false)).toEqual(0);
-      });
-
-      it('should return -1 if isWrapping is true and start is less than end', () => {
-        const isWrapping = true;
-        const start = 0;
-        const end = 6;
-
-        expect(instance.getSlideDirection(start, end, isWrapping)).toEqual(-1);
-      });
-
-      it('should return 1 if isWrapping is true and start is greater than end', () => {
-        const isWrapping = true;
-        const start = 6;
-        const end = 0;
-
-        expect(instance.getSlideDirection(start, end, isWrapping)).toEqual(1);
-      });
-
-      it('should return 1 if isWrapping is false and start is less than end', () => {
-        const isWrapping = false;
-        const start = 0;
-        const end = 6;
-
-        expect(instance.getSlideDirection(start, end, isWrapping)).toEqual(1);
-      });
-
-      it('should return -1 if isWrapping is false and start is greater than end', () => {
-        const isWrapping = false;
-        const start = 6;
-        const end = 0;
-
-        expect(instance.getSlideDirection(start, end, isWrapping)).toEqual(-1);
-      });
     });
   });
